refactor: useRecordFormState로 form 상태 분리
